LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/stroke-expected.txt


FAIL 'stroke' does not support 'black' assert_class_string: Unsupported value must be a CSSStyleValue and not one of its subclasses expected "[object CSSStyleValue]" but got "[object CSSKeywordValue]"
FAIL 'stroke' does not support 'red gray' assert_not_equals: Unsupported value must not be null got disallowed value null
